back/function-event-management/requirements.txt
azure-functions
pyodbc
requests_toolbelt
